+++ StringView.h
@@ -179,14 +179,14 @@ namespace neo
                 ++current;
                 if (*current == by)
                 {
-                    strings.construct(_begin.ptr(), current.ptr() - _begin.ptr());
+                    strings.construct(_begin.ptr(), static_cast<size_t>(current.ptr() - _begin.ptr()));
                     while (*current == by)
                         ++current;
                     _begin = current;
                 }
             } while (current != _end);
             if (_begin != _end)
-                strings.construct(_begin.ptr(), current.ptr() - _begin.ptr());
+                strings.construct(_begin.ptr(), static_cast<size_t>(current.ptr() - _begin.ptr()));
             return strings;
         }
 
@@ -202,7 +202,7 @@ namespace neo
                 ++current;
                 if (StringView(current.ptr(), min(by.byte_size(), (size_t)(_end.ptr() - current.ptr()))).starts_with(by))
                 {
-                    strings.construct(_begin.ptr(), current.ptr() - _begin.ptr());
+                    strings.construct(_begin.ptr(),static_cast<size_t>(current.ptr() - _begin.ptr()));
                     do
                     {
                         for (auto to_skip = by.length(); to_skip > 0; to_skip--)
@@ -212,7 +212,7 @@ namespace neo
                 }
             } while (current != _end);
             if (_begin != _end)
-                strings.construct(_begin.ptr(), current.ptr() - _begin.ptr());
+                strings.construct(_begin.ptr(), static_cast<size_t>(current.ptr() - _begin.ptr()));
             return strings;
         }
